+++ org-code-javabuilder/media/src/main/java/org/code/media/support/AudioWriter.java
@@ -1,8 +1,10 @@
-package org.code.media;
+package org.code.media.support;
 
 import java.io.ByteArrayOutputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import org.code.media.SoundException;
+import org.code.media.util.AudioUtils;
 import org.code.protocol.InternalErrorKey;
 import org.code.protocol.InternalServerRuntimeError;
