+++ org-code-javabuilder/media/src/main/java/org/code/media/support/FontHelper.java
@@ -1,7 +1,8 @@
-package org.code.media;
+package org.code.media.support;
 
 import java.awt.*;
 import java.util.Map;
+import org.code.media.FontStyle;
 import org.code.protocol.CachedResources;
 
 public class FontHelper {
@@ -11,19 +12,19 @@ public class FontHelper {
     this.fontMap = CachedResources.getInstance().getFontMap();
   }
 
-  public java.awt.Font getFont(Font font, FontStyle fontStyle) {
+  public java.awt.Font getFont(org.code.media.Font font, FontStyle fontStyle) {
     return this.fontMap.get(getFontFilename(font, fontStyle));
   }
 
-  private String getFontFilename(Font font, FontStyle fontStyle) {
+  private String getFontFilename(org.code.media.Font font, FontStyle fontStyle) {
     return fontToFilenamePrefix.get(font) + fontStyleToFilenameSuffix.get(fontStyle);
   }
 
-  private static Map<Font, String> fontToFilenamePrefix =
+  private static Map<org.code.media.Font, String> fontToFilenamePrefix =
       Map.ofEntries(
-          Map.entry(Font.MONO, "LiberationMono"),
-          Map.entry(Font.SANS, "LiberationSans"),
-          Map.entry(Font.SERIF, "LiberationSerif"));
+          Map.entry(org.code.media.Font.MONO, "LiberationMono"),
+          Map.entry(org.code.media.Font.SANS, "LiberationSans"),
+          Map.entry(org.code.media.Font.SERIF, "LiberationSerif"));
 
   private static Map<FontStyle, String> fontStyleToFilenameSuffix =
       Map.ofEntries(
