+++ org-code-javabuilder/neighborhood/src/main/java/org/code/neighborhood/support/Direction.java
@@ -1,4 +1,4 @@
-package org.code.neighborhood;
+package org.code.neighborhood.support;
 
 public enum Direction {
   NORTH("north"),
