+++ org-code-javabuilder/neighborhood/src/main/java/org/code/neighborhood/support/GridFactory.java
@@ -1,4 +1,4 @@
-package org.code.neighborhood;
+package org.code.neighborhood.support;
 
 import java.io.File;
 import java.io.FileInputStream;
