+++ org-code-javabuilder/neighborhood/src/main/java/org/code/neighborhood/support/NeighborhoodSignalMessage.java
@@ -1,4 +1,4 @@
-package org.code.neighborhood;
+package org.code.neighborhood.support;
 
 import java.util.HashMap;
 import org.code.protocol.ClientMessage;
