+++ org-code-javabuilder/neighborhood/src/main/java/org/code/neighborhood/support/World.java
@@ -1,4 +1,4 @@
-package org.code.neighborhood;
+package org.code.neighborhood.support;
 
 import java.io.IOException;
 import org.code.protocol.*;
