+++ org-code-javabuilder/neighborhood/src/test/java/org/code/neighborhood/PainterTest.java
@@ -5,6 +5,7 @@ import static org.mockito.Mockito.*;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
+import org.code.neighborhood.support.*;
 import org.code.protocol.*;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
