+++ org-code-javabuilder/neighborhood/src/test/java/org/code/neighborhood/support/GridTest.java
@@ -1,4 +1,4 @@
-package org.code.neighborhood;
+package org.code.neighborhood.support;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.mock;
