+++ org-code-javabuilder/theater/src/test/java/org/code/theater/StageTest.java
@@ -10,11 +10,11 @@ import java.awt.image.BufferedImage;
 import java.io.ByteArrayOutputStream;
 import java.io.FileNotFoundException;
 import java.io.PrintStream;
-import org.code.media.AudioWriter;
 import org.code.media.Color;
 import org.code.media.Font;
 import org.code.media.FontStyle;
 import org.code.media.Image;
+import org.code.media.support.AudioWriter;
 import org.code.protocol.*;
 import org.code.theater.support.*;
 import org.junit.jupiter.api.AfterEach;
