+++ org-code-javabuilder/validation/src/main/java/org/code/validation/support/NeighborhoodTracker.java
@@ -7,8 +7,8 @@ import java.util.*;
 import java.util.function.Function;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
-import org.code.neighborhood.NeighborhoodSignalKey;
-import org.code.neighborhood.World;
+import org.code.neighborhood.support.NeighborhoodSignalKey;
+import org.code.neighborhood.support.World;
 import org.code.protocol.ClientMessage;
 import org.code.protocol.ClientMessageType;
 import org.code.validation.NeighborhoodLog;
